Move DotRouter into own file
